without registration, with all databases, but don't work user review update, another all work

+++ src/main/java/com/iba/YourReview/repository/AppUserRepo.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AppUserRepo extends CrudRepository<AppUser, Integer> {
+public interface AppUserRepo extends CrudRepository<AppUser, Long> {
 
     AppUser findByNickName(String nickName);
